📝 chore(tailwind.config.cjs): add tailwind configuration file
The tailwind.config.cjs file is added to the project. This file contains the configuration for Tailwind CSS. It specifies the content to be processed by Tailwind CSS, extends the default theme with customizations, and adds the 'manrope' font family to the sans font stack. Additionally, the '@tailwindcss/typography' plugin is included.